requirements.txt
llama-index>=0.8.53
python-dotenv>=1.0.0
